+++ examples/misc/feature_misc/lib/lifecycle_handlers/di_registration_lifecycle_handler.g.dart
@@ -1,0 +1,11 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'di_registration_lifecycle_handler.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+DIRegistrationLifecycleHandler _$DIRegistrationLifecycleHandlerFromJson(
+        Map<String, dynamic> json) =>
+    DIRegistrationLifecycleHandler();